Make FeaturesView component generic
