+++ src/tools/blast/utils/__tests__/hsp.spec.ts
@@ -4,7 +4,7 @@ import {
   getOffset,
 } from '../../../utils/sequences';
 import mockData from '../__mocks__/hspMocks.json';
-import { FeatureData } from '../../../../uniprotkb/components/protein-data-views/FeaturesView';
+import { FeatureData } from '../../../../uniprotkb/components/protein-data-views/UniProtKBFeaturesView';
 
 describe.skip('HSP util tests', () => {
   it('should work with longer query than hit', () => {
