+++ src/tools/utils/feature.ts
@@ -1,4 +1,4 @@
-import { ProcessedFeature } from '../../uniprotkb/components/protein-data-views/FeaturesView';
+import { ProcessedFeature } from '../../shared/components/views/FeaturesView';
 import { processUrlTemplate } from '../../uniprotkb/components/protein-data-views/XRefView';
 import evidenceUrls from '../../uniprotkb/config/evidenceUrls';
 import FeatureType from '../../uniprotkb/types/featureType';
