+++ src/tools/utils/sequences.ts
@@ -1,11 +1,11 @@
 /* eslint-disable camelcase */
 import { getColorByType } from 'protvista-track';
 import { MSAInput } from '../components/AlignmentView';
+import { FeatureData } from '../../uniprotkb/components/protein-data-views/UniProtKBFeaturesView';
 import {
-  ProcessedFeature,
-  FeatureData,
   LocationModifier,
-} from '../../uniprotkb/components/protein-data-views/FeaturesView';
+  ProcessedFeature,
+} from '../../shared/components/views/FeaturesView';
 
 export const findSequenceSegments = (seq: string) => {
   const ranges: number[][] = [];
