+++ src/uniparc/components/entry/UniParcFeaturesView.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+import { SequenceFeature } from '../../adapters/uniParcConverter';
+
+const UniParcFeaturesView: FC<{ data: SequenceFeature[] }> = ({ data }) => null;
+export default UniParcFeaturesView;
